First pass at csv support
Due to Fred freaking out about me using hard coded data I decided I
would try to show him whats good and add csv support. The problem is the
excel sheet we're saving everything is not designed for being read by
machines and needs a lot of cleanup. It is also missing a lot of data
that I had added to my previously hardcoded data and I'm just not
willing to put this much effort into this right now.